Move web-only files to platform/web/ (Phase 1) (#786)
* Move web-only files to platform/web/ (Phase 1)

Move files whose consumers are entirely within web code to
src/platform/web/: compatibility.ts, addLibsamplerateModule.ts,
input.ts (MediaDeviceInput), output.ts (MediaDeviceOutput).

Split VoiceSessionSetup into common (types + injectable strategy
variable) and web (implementation + side-effect registration).

Extract shared types (InputConfig, InputEventTarget, OutputConfig,
PlaybackEventTarget, etc.) into common InputController.ts and
OutputController.ts so they remain accessible without DOM.

Refactor applyDelay to accept a resolved delay value directly,
removing the dependency on web-only compatibility.ts.

Files still depending on common code (rawAudioProcessor,
scribeAudioProcessor, createWorkletModuleLoader,
createAnalyserVolumeProvider) stay in src/utils/ until their
injection refactoring in Phases 5-6.


* Remove unused type re-exports from web input/output

Types are defined in common InputController.ts / OutputController.ts
and imported from there by all consumers. The re-exports from the
web files were dead code.


* Fix IIFE bundle entry point to include web strategy

The IIFE build (for CDN/script-tag users) used src/index.ts as entry,
which no longer imports the web session setup strategy after the
platform split. Change entry to src/platform/web/index.ts which
re-exports everything plus the side-effect registration.

Also adds a TODO for the lost Android delay default in resolveDelay.


* Call assertRuntimeCompatibility() in Conversation.startSession

Validates that required cross-platform globals (fetch, WebSocket,
TextEncoder, etc.) are available before starting a session, giving
users a clear error message instead of cryptic failures.


---------
