+++ packages/client/package.json
@@ -17,7 +17,7 @@
     "generate-version": "printf \"// This file is auto-generated during build\\nexport const PACKAGE_VERSION = \\\"%s\\\";\\n\" \"$npm_package_version\" > src/version.ts",
     "generate-worklets": "node scripts/generateWorklets.js",
     "build:esm": "tsc --build",
-    "build:iife": "rolldown src/index.ts -f iife --name ElevenLabsClient -o dist/lib.iife.js --sourcemap --external react --external react-dom",
+    "build:iife": "rolldown src/platform/web/index.ts -f iife --name ElevenLabsClient -o dist/lib.iife.js --sourcemap --external react --external react-dom",
     "build": "node --run build:esm && node --run build:iife",
     "dev": "tsc --build --watch",
     "clean": "tsc --build --clean && rm -rf ./dist",
