+++ packages/client/src/index.ts
@@ -1,4 +1,5 @@
 import { isTextOnly, type PartialOptions } from "./BaseConversation.js";
+import { assertRuntimeCompatibility } from "./runtime.js";
 import { TextConversation } from "./TextConversation.js";
 import { VoiceConversation } from "./VoiceConversation.js";
 
@@ -22,8 +23,8 @@ export type {
   OutputController,
   OutputDeviceConfig,
 } from "./OutputController.js";
-export type { InputConfig } from "./utils/input.js";
-export type { OutputConfig } from "./utils/output.js";
+export type { InputConfig } from "./InputController.js";
+export type { OutputConfig } from "./OutputController.js";
 export type {
   IncomingSocketEvent,
   VadScoreEvent,
@@ -91,6 +92,7 @@ interface ConversationNamespace {
 
 export const Conversation: ConversationNamespace = {
   startSession(options: PartialOptions) {
+    assertRuntimeCompatibility();
     return isTextOnly(options)
       ? TextConversation.startSession(options)
       : VoiceConversation.startSession(options);
