+++ packages/client/src/platform/web/compatibility.ts
