+++ packages/client/src/platform/web/index.ts
@@ -1,1 +1,4 @@
+// Side-effect: registers the web voice session setup strategy
+import "./VoiceSessionSetup.js";
+
 export * from "../../index.js";
