+++ packages/client/src/utils/input.test.ts
@@ -1,5 +1,6 @@
 import { describe, it, expect, afterEach } from "vitest";
-import { MediaDeviceInput, type InputMessageEvent } from "./input.js";
+import { MediaDeviceInput } from "../platform/web/input.js";
+import type { InputMessageEvent } from "../InputController.js";
 
 describe("MediaDeviceInput", () => {
   let input: MediaDeviceInput | null = null;
